% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_tcga2018_data}
\alias{all_tcga2018_data}
\title{Data from 2018 TCGA studies from cBioPortal}
\format{
A data frame with 14944 rows and 6 variables:
\describe{
    \item{hugoGeneSymbol}{hugo gene symbol}
    \item{Gain}{proportion of cohort with gain in this gene}
    \item{Amplification}{proportion of cohort with amplification in this gene}
    \item{ShallowDeletion}{proportion of cohort with shallow deletion in this gene}
    \item{DeepDeletion}{proportion of cohort with deep deletion in this gene}
    \item{study_name}{cancer type and sample size}
    }
}
\source{
\url{https://github.com/waldronlab/cBioPortalData} See data-raw folder.
}
\usage{
all_tcga2018_data
}
\description{
A dataset containing the study name and aggregated gene level copy number data
}
\keyword{datasets}
