% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbio_studies}
\alias{cbio_studies}
\title{Names of 2018 TCGA studies from cBioPortal}
\format{
A data frame with 32 rows and 2 variables:
\describe{
    \item{Cancer}{Name of diagnosis and sample size}
    \item{studyId}{studyId that can be used in the cBioPortalData R package}
    }
}
\source{
\url{https://github.com/waldronlab/cBioPortalData} See data-raw folder.
}
\usage{
cbio_studies
}
\description{
A dataset containing the names and studyIds of the 2018 TCGA studies from cBioPortal.
}
\keyword{datasets}
