% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytoband_data}
\alias{cytoband_data}
\title{Genomic locations of cytoband labels}
\format{
A data frame with 863 rows and 6 variables:
\describe{
    \item{chrom}{chromosome}
    \item{chromStart}{start position}
    \item{chromEnd}{end position}
    \item{name}{cytoband name}
    \item{gieStain}{color}
    \item{color}{HEX color}
}
}
\source{
\url{https://genome.ucsc.edu/cgi-bin/hgTables}
}
\usage{
cytoband_data
}
\description{
A dataset containing the chr, start and end position for cytobands
according to hg38.
}
\keyword{datasets}
