% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_data}
\alias{gene_data}
\title{Gene data for vignette example}
\format{
A dataframe with 112 rows and 6 variables
\describe{
    \item{chr}{chromosome}
    \item{start}{start location}
    \item{end}{end location}
    \item{gene}{gene name}
    \item{log2}{log2 copy number ratio}
    \item{weight}{weight given to log2 value}
    \item{loh}{loss of heterozygosity}
    }
}
\source{
Center for Personalized Diagnositics at the University of Pennsylvania
}
\usage{
data(gene_data)
}
\description{
A dataset containing simulated gene data as sample input for launchCNViz
}
\keyword{datasets}
