% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_data}
\alias{meta_data}
\title{Metadata for vignette example}
\format{
A dataframe with 1 rows and 2 variables
\describe{
    \item{purity}{sample purity}
    \item{ploidy}{tumor ploidy}
    }
}
\source{
Center for Personalized Diagnositics at the University of Pennsylvania
}
\usage{
data(meta_data)
}
\description{
A dataset containing simulated metadata as sample input for launchCNViz
}
\keyword{datasets}
