% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{probe_data}
\alias{probe_data}
\title{Probe data for vignette example}
\format{
A data frame with 2006 rows and 6 variables:
\describe{
    \item{chr}{chromosome}
    \item{start}{start location}
    \item{end}{end location}
    \item{gene}{gene name}
    \item{log2}{log2 copy number ratio}
    \item{weight}{weight given to log2 value}
    }
}
\source{
Center for Personalized Diagnositics at the University of Pennsylvania
}
\usage{
data(probe_data)
}
\description{
A dataset containing simulated probe data as sample input for launchCNViz
}
\keyword{datasets}
