% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{segment_data}
\alias{segment_data}
\title{Segment data for vignette example}
\format{
A dataframe with 101 rows and 5 variables
\describe{
    \item{chr}{chromosome}
    \item{start}{start location}
    \item{end}{end location}
    \item{log2}{log2 copy number ratio}
    \item{loh}{loss of heterozygosity}
    }
}
\source{
Center for Personalized Diagnositics at the University of Pennsylvania
}
\usage{
data(segment_data)
}
\description{
A dataset containing simulated segment data as sample input for launchCNViz
}
\keyword{datasets}
