% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variant_data}
\alias{variant_data}
\title{Variant data for vignette example}
\format{
A dataframe with 119 rows and 4 variables
\describe{
    \item{gene}{gene name}
    \item{mutation_id}{string with information about snv}
    \item{depth}{read depth}
    \item{start}{starting location}
    }
}
\source{
Center for Personalized Diagnositics at the University of Pennsylvania
}
\usage{
data(variant_data)
}
\description{
A dataset containing simulated SNV and indel data as sample input for launchCNViz
}
\keyword{datasets}
