% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchCNViz.R
\name{launchCNViz}
\alias{launchCNViz}
\title{Launches CNViz, a shiny app to visualize your sample's copy number data.}
\usage{
launchCNViz(
  sample_name = "sample",
  probe_data = data.frame(),
  gene_data = data.frame(),
  segment_data = data.frame(),
  snv_data = data.frame(),
  meta_data = data.frame()
)
}
\arguments{
\item{sample_name}{A string with the ID of your sample.}

\item{probe_data}{A dataframe containing probe-level data. Column names must include chr, gene, start, end, log2. Optional column: weight.}

\item{gene_data}{A dataframe containing gene-level data. Column names must include chr, gene, start, end, log2. Optional columns: weight, loh; where loh values are TRUE or FALSE}

\item{segment_data}{A dataframe containing segment-level data. Column names must include chr, start, end, log2. Optional column: loh; where loh values are TRUE or FALSE}

\item{snv_data}{A dataframe containg SNVs and columns of your choosing. The only required column is gene. Optional columns: start, mutation_id; where start indicates the starting position of the mutation and mutation_id is a string in any format. Additional columns might include depth, allelic_fraction.}

\item{meta_data}{A dataframe containing your sample's metadata - columns of your choosing. Columns might include purity and ploidy. Ploidy will be rounded to the nearest whole number. This dataframe should only have one row.}
}
\description{
CNViz launches a shiny application to visualize your sample's copy number data.
At least one of probe_data, gene_data, or segment_data must be supplied;
sample_name, meta_data, snv and karyotype_pdf are all optional.
}
